+++ NEWS.md
@@ -1,7 +1,8 @@
 # solvebio 2.12.0
 
 * Add support for enabling/disabling Global Beacon on datasets
-* Add ability to perform Global Search
+* Add ability to perform Global Search queries
+
 
 # solvebio 2.11.0